requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
apturl==0.5.2
attrs==24.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-mgmt-compute==30.5.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.3.0
azure-mgmt-resource==23.0.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
ccsm==0.9.14.1
certifi==2024.7.4
cffi==1.17.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
cryptography==43.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.14.1
future==0.18.2
gpg==1.16.0
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
install==1.3.5
isodate==0.6.1
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msal==1.27.0
msal-extensions==1.1.0
netifaces==0.11.0
numpy==2.0.1
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.5.0
portalocker==2.8.2
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
schedule==1.2.2
SecretStorage==3.3.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
tomli==2.0.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
